requirement1s.txt
absl-py==1.1.0
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
arabic-reshaper==2.1.3
asgiref==3.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
citeproc-py==0.6.0
citeproc-py-styles==0.1.3
click==8.1.3
colorama==0.4.4
cryptography==3.4.6
cssselect2==0.6.0
Cython==0.29.30
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.1.7
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-photologue==3.13
django-rest-framework==0.1.0
django-sortedm2m==3.0.2
django-vitae==0.1.0
django-widgets==0.1.16
djangorestframework==3.12.2
djangorestframework-recursive==0.1.2
docutils==0.18.1
ExifRead==2.1.2
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
grpcio==1.46.3
gunicorn==20.1.0
html5lib==1.1
idna==3.1
importlib-metadata==4.11.4
jsonfield==3.1.0
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
lxml==4.8.0
Markdown==3.3.7
multidict==6.0.2
numpy==1.22.4
oauthlib==3.1.0
oscrypto==1.3.0
Pillow==8.1.2
protobuf==4.21.1
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.12.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.0.1
PyPDF3==1.0.6
pypiwin32==223
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
qrcode==7.3.1
recursive==1.3.0
reportlab==3.6.9
requests==2.28.0
requests-oauthlib==1.3.0
scipy==1.8.1
six==1.16.0
smart-open==6.0.0
soupsieve==2.2.1
SQLAlchemy==1.4.37
sqlalchemy-mptt==0.2.5
sqlparse==0.4.1
svglib==1.3.0
tensorboard==1.15.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==6.2.0
xhtml2pdf==0.2.7
yarl==1.7.2
zipp==3.8.0
